Added DI Annotations
