+++ pom.xml
@@ -30,7 +30,13 @@
             <version>4.0.1</version>
             <scope>provided</scope>
         </dependency>
-      </dependencies>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>5.2.9.RELEASE</version>
+        </dependency>
+
+    </dependencies>
 
     <build>
         <plugins>
