+++ src/main/java/com/example/spring_servletcontainer/repository/PostRepository.java
@@ -1,48 +1,14 @@
 package com.example.spring_servletcontainer.repository;
 
-import java.util.*;
-import java.util.concurrent.ArrayBlockingQueue;
+import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.atomic.AtomicInteger;
 
-public class PostRepository {
-    private final int BLOCKING_QUEUE_CAPACITY = 255;
-    private final int INDEX_INITIAL_VALUE = 0;
-    BlockingQueue<Post> posts = new ArrayBlockingQueue<>(BLOCKING_QUEUE_CAPACITY);
-    private final AtomicInteger index = new AtomicInteger(INDEX_INITIAL_VALUE);
+public interface PostRepository {
+    BlockingQueue<Post> all();
 
-    private int autoincrement() {
-        return index.incrementAndGet();
-    }
+    Optional<Post> getById(long id);
 
-    public BlockingQueue<Post> all() {
-        return posts;
-    }
+    Post save(Post post);
 
-    public Optional<Post> getById(long id) {
-        return posts.stream().filter(x -> x.getId() == id).findFirst();
-    }
-
-    public Post save(Post post) {
-        //Ниже обработка когда id = 0
-
-        if (post.getId() == 0) {
-            post.setId(autoincrement());
-            posts.add(post);
-        } else {
-            //Ниже обработка когда id != 0
-            Optional<Post> postOptional = getById(post.getId());
-            if (postOptional.isPresent()) {
-                Post updatedPost = postOptional.get();
-                posts.remove(updatedPost);
-                updatedPost.setMessage(post.getMessage());
-                posts.add(updatedPost);
-            }
-        }
-        return post;
-    }
-
-    public void removeById(long id) {
-        posts.removeIf(x -> x.getId() == id);
-    }
+    void removeById(long id);
 }
