+++ src/main/java/com/example/spring_servletcontainer/service/PostService.java
@@ -1,15 +1,20 @@
 package com.example.spring_servletcontainer.service;
 
 import com.example.spring_servletcontainer.repository.Post;
-import com.example.spring_servletcontainer.repository.PostRepository;
+import com.example.spring_servletcontainer.repository.PostRepositoryImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
 
+@Service
 public class PostService {
-    PostRepository repository;
 
-    public PostService(PostRepository repository) {
+    PostRepositoryImpl repository;
+
+    @Autowired
+    public PostService(PostRepositoryImpl repository) {
         this.repository = repository;
     }
